requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
easyocr
requests
jinja2
pytest
